+++ src/corva/configuration.py
@@ -18,12 +18,5 @@ class Settings(pydantic.BaseSettings):
     APP_KEY: str  # <provider-name-with-dashes>.<app-name-with-dashes>
     PROVIDER: str
 
-    @property
-    def APP_NAME(self) -> str:
-        app_name_with_dashes = self.APP_KEY.split('.')[1]
-        app_name = app_name_with_dashes.replace('-', ' ').title()
-
-        return app_name
-
 
 SETTINGS = Settings()