+++ src/corva/models/scheduled.py
@@ -5,7 +5,8 @@ from typing import List
 
 import pydantic
 
-from corva.models.base import BaseContext, CorvaBaseEvent, RawBaseEvent
+from corva.api import Api
+from corva.models.base import CorvaBaseEvent, RawBaseEvent
 
 
 class ScheduledEvent(CorvaBaseEvent):
@@ -78,6 +79,6 @@ class RawScheduledEvent(CorvaBaseEvent, RawBaseEvent):
 
         return events
 
-
-class ScheduledContext(BaseContext[RawScheduledEvent]):
-    pass
+    def set_schedule_as_completed(self, api: Api) -> None:
+        """Sets schedule as completed."""
+        api.post(path=f'scheduler/{self.schedule_id}/completed')