+++ src/corva/models/task.py
@@ -1,9 +1,19 @@
 from __future__ import annotations
 
+import enum
+from typing import List
+
 import pydantic
+import requests
 from pydantic.types import conint
 
-from corva.models.base import BaseContext, CorvaBaseEvent, RawBaseEvent
+from corva.api import Api
+from corva.models.base import CorvaBaseEvent, RawBaseEvent
+
+
+class TaskStatus(enum.Enum):
+    fail = 'fail'
+    success = 'success'
 
 
 class TaskEvent(CorvaBaseEvent):
@@ -25,9 +35,18 @@ class RawTaskEvent(CorvaBaseEvent, RawBaseEvent):
     version: conint(ge=2, le=2)  # only utils API v2 supported
 
     @staticmethod
-    def from_raw_event(event: dict) -> RawTaskEvent:
-        return pydantic.parse_obj_as(RawTaskEvent, event)
+    def from_raw_event(event: dict) -> List[RawTaskEvent]:
+        return [pydantic.parse_obj_as(RawTaskEvent, event)]
 
+    def get_task_event(self, api: Api) -> TaskEvent:
+        response = api.get(path=f'v2/tasks/{self.task_id}')
+        response.raise_for_status()
 
-class TaskContext(BaseContext[RawTaskEvent]):
-    pass
+        return TaskEvent(**response.json())
+
+    def update_task_data(
+        self, api: Api, status: TaskStatus, data: dict
+    ) -> requests.Response:
+        """Updates the task."""
+
+        return api.put(path=f'v2/tasks/{self.task_id}/{status.value}', data=data)